openpgp-card-sequoia: Allow disabling default crypto backend.
Previously the list of features of `sequoia-openpgp` unconditionally
included `crypto-nettle`. That was preventing the use of this crate in
pure Rust setups.

This commit allows disabling crypto backends and it is up to the
crate user to explicitly include `sequoia_openpgp` features that they
want.

See: https://gitlab.com/sequoia-pgp/sequoia/-/blob/main/README.md#note